+ Исправление замечаний code review №3
