--- cmd/staticlint/config.json
@@ -1,19 +1,0 @@
-{
-  "staticcheck": [
-    "S1001",
-    "S1002",
-    "S1025",
-    "S1028",
-    "ST1000",
-    "ST1005",
-    "ST1006",
-    "ST1017",
-    "ST1020",
-    "ST1021",
-    "ST1022",
-    "ST1023",
-    "QF1003",
-    "QF1006",
-    "QF1007"
-  ]
-}
