+++ cmd/staticlint/config/config.go
@@ -1,0 +1,8 @@
+package config
+
+// ConfigData структура с конфигурацией анализаторов пакета staticcheck
+//
+// SA добавлены полностью. Для добавления дополнительных прописать необходимый в файле `config.json`
+type ConfigData struct {
+	Staticcheck []string
+}
