+++ cmd/staticlint/config/config_file.go
@@ -1,6 +1,6 @@
 //go:build config
 
-package main
+package config
 
 import (
 	"embed"
