+++ cmd/staticlint/main_test.go
